+++ conreality_beacon/lib/src/beacon_event.dart
@@ -1,0 +1,6 @@
+/* This is free and unencumbered software released into the public domain. */
+
+/// Represents a beacon event.
+abstract class BeaconEvent {
+  const BeaconEvent();
+}