+++ conreality_beacon/CHANGELOG.md
@@ -6,6 +6,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Upcoming release
+### Added
+- `Beacons.scan()` static method
+- `Beacon` class
+- `BeaconEvent` class
+- `BeaconProtocol` enum
+- `BeaconScan` class
+- AndroidX `@NonNull`/`@Nullable`/etc annotations
+- [AndroidX lifecycle](https://developer.android.com/topic/libraries/architecture/lifecycle) support
+
 ## 0.0.1 - 2019-07-13
 ### Added
 - `ConrealityBeacon.version` getter